Implement Versioning and infoV1
