+++ src/modules/person/api/personinfo/v1/person-info.response.v1.ts
@@ -1,0 +1,15 @@
+/* eslint-disable max-classes-per-file */
+import { ApiProperty } from '@nestjs/swagger';
+import { PersonInfoResponse, PersonNestedInPersonInfoResponse } from '../person-info.response.js';
+
+class PersonNestedInPersonInfoResponseV1 extends PersonNestedInPersonInfoResponse {}
+
+export class PersonInfoResponseV1 extends PersonInfoResponse {
+    @ApiProperty()
+    public override person: PersonNestedInPersonInfoResponseV1;
+
+    public constructor(props: Readonly<PersonInfoResponseV1>) {
+        super(props);
+        this.person = new PersonNestedInPersonInfoResponseV1(props.person);
+    }
+}
